Build: (c2c2dd2) Go binary instead of uwsgi config.

+++ static/install.sh
@@ -1,28 +1,45 @@
-#!/usr/bin/bash
-
-TARGET_DIR=/etc/flingdev
-SITE_PACKAGES_FOLDER=`python -c "import sysconfig; print(sysconfig.get_path('purelib'))"`
+#!/bin/bash
 
 echo "Installing loophost..."
-sudo mkdir -p $TARGET_DIR/certs && chown -R joshuamckenty $TARGET_DIR
-pushd $TARGET_DIR
+
+TARGET_DIR=~/.flingdev
+
 pip install fling-cli fling-hub uwsgi
+HUBDIR=`python -c "import os; import fling_hub; print(os.path.dirname(os.path.realpath(fling_hub.__file__)))"`
+
+mkdir -p $TARGET_DIR/certs
+pushd $TARGET_DIR
+CWD=`pwd`
 
 fling auth
+if [ ! -f flinguser.txt ]
+then
+    echo "Authentication failed... loophost setup aborted."
+    exit 1
+fi
 # Save the github username into apps.json
-USERNAME=joshuamckenty
+USERNAME=$(<flinguser.txt)
 
-certbot certonly --authenticator fling-authenticator -d "*.$USERNAME.fling.dev" \
+certbot certonly --non-interactive --expand --agree-tos -m webmaster@fling.dev \
+     --authenticator fling_authenticator -d "*.$USERNAME.fling.dev"  -d "$USERNAME.fling.dev"\
     --config-dir . --work-dir . --logs-dir . --fling_authenticator-propagation-seconds 30
 cp "./live/$USERNAME.fling.dev/fullchain.pem" ./certs/$USERNAME.fling.dev.crt
 cp "./live/$USERNAME.fling.dev/privkey.pem" ./certs/$USERNAME.fling.dev.key
 
-cp $SITE_PACKAGES_FOLDER/fling_hub/hubconfig.ini ./
-# cp $SITE_PACKAGES_FOLDER/fling_hub/flingroute.py ./
-# cp $SITE_PACKAGES_FOLDER/fling_hub/logo-hc.txt ./
-# TODO(don't clobber if it exists)!
+cp $HUBDIR/hubconfig.ini ./
 
-echo "{\"apps\": {}, \"fqdn\": \"$USERNAME.fling.dev\"}" > loophost.json
+if [ ! -f loophost.json ]
+then 
+    echo "{\"apps\": {}, \"fqdn\": \"$USERNAME.fling.dev\"}" > loophost.json
+fi
 
-sudo cp $SITE_PACKAGES_FOLDER/fling_hub/hub.plist /Library/LaunchDaemons/
+# TODO(DEAL WITH MULTIPLE USERS)
+cp $HUBDIR/loophost.plist.template ./
+cp $HUBDIR/hub.plist.template ./
+sed -e "s@\${USERNAME}@$USERNAME@" -e "s@\${CWD}@$CWD@" -e "s@\${HUBDIR}@$HUBDIR@" loophost.plist.template | sudo tee /Library/LaunchDaemons/dev.fling.hub.local.plist
+sed -e "s@\${USERNAME}@$USERNAME@" -e "s@\${CWD}@$CWD@" -e "s@\${HUBDIR}@$HUBDIR@" hub.plist.template | sudo tee /Library/LaunchDaemons/dev.fling.hub.plist
+sudo launchctl load /Library/LaunchDaemons/dev.fling.hub.local.plist
+sudo launchctl load /Library/LaunchDaemons/dev.fling.hub.plist
+
+
 popd
